update signer repo
