+++ cli/src/updater/github.rs
@@ -14,8 +14,7 @@ pub(crate) async fn fetch_release_runtimes(
     let release = octocrab::instance()
         .repos(&config.owner, &config.repo)
         .releases()
-        .get_by_tag("v0.9.23")
-        // .get_latest()
+        .get_latest()
         .await?;
     info!("📅 Latest release: {}", release.tag_name);
     for asset in release.assets {
