+++ src/components/QrCode.tsx
@@ -1,4 +1,6 @@
 import {QrInfo, RpcSource, WasmSource} from "../scheme";
+import Hash from "./Hash";
+import {row} from "./Specs";
 
 export default function QrCode({ path, source }: QrInfo) {
   let source_block = null;
@@ -7,22 +9,19 @@ export default function QrCode({ path, source }: QrInfo) {
       case "Wasm": {
         const s = source as WasmSource;
         source_block = (
-            <div>
-              Source:
-              Repo: {s.github_repo}
-              Hash: {s.hash}
-            </div>
+            <ul className="flex">
+              {row("Metadata source", <a href={`https://github.com/${s.github_repo}/releases`} target="_blank" rel="noreferrer">{s.github_repo}</a>)}
+              {row("Blake2-256 hash", <Hash value={s.hash} />)}
+            </ul>
         )
         break
       }
       case "Rpc": {
         const s = source as RpcSource;
         source_block = (
-            <div>
-              <p>{"Source:"}</p>
-              <p>{"Url: "} {s.url}</p>
-              <p>{"Block: "} {s.block}</p>
-            </div>
+            <ul className="flex">
+              {row("Source block", <Hash value={s.block} />, true)}
+            </ul>
         )
         break
       }
