+++ src/components/Specs.tsx
@@ -1,48 +1,15 @@
-import { ReactElement, useEffect, useState } from "react";
+import { ReactElement } from "react";
 import { ChainSpec } from "../scheme";
 import "./Specs.css";
-import Copy from "../assets/copy.svg";
+import Hash from "./Hash";
 
 interface SpecsProps {
   chainSpecs: ChainSpec;
   color?: string;
 }
 
-export const copyToClipboard = (text: string): void => {
-  const dummy = document.createElement("textarea");
-  document.body.appendChild(dummy);
-  dummy.value = text;
-  dummy.select();
-  document.execCommand("copy");
-  document.body.removeChild(dummy);
-};
-
 export default function Specs({ chainSpecs, color }: SpecsProps) {
   const { rpcEndpoint, genesisHash, unit, base58prefix } = chainSpecs;
-  const [copied, setCopied] = useState<boolean>(false);
-
-  useEffect(() => {
-    setTimeout(() => copied && setCopied(false), 2000);
-  }, [copied]);
-
-  const elipsisHash = (el: string) => {
-    const sliced = el.slice(0, 6) + "..." + el.slice(el.length - 4, el.length);
-    const cName = copied ? "fade" : "hidden";
-    return (
-      <>
-        <div className="flex float-left">{sliced}</div>
-        <img
-          className="w-5 h-5 ml-2 cursor-pointer"
-          src={Copy}
-          onClick={() => {
-            setCopied(true);
-            copyToClipboard(el);
-          }}
-        />
-        <div className={"text-green-500 ml-3 ".concat(cName)}>Copied</div>
-      </>
-    );
-  };
 
   return (
     <>
@@ -56,7 +23,7 @@ export default function Specs({ chainSpecs, color }: SpecsProps) {
         )}
       </ul>
       <ul className="flex">
-        {row("Genesis hash", elipsisHash(genesisHash))}
+        {row("Genesis hash", <Hash value={genesisHash} />  )}
         {row("Address prefix", base58prefix.toString())}
       </ul>
       <ul className="flex">
@@ -76,7 +43,7 @@ export default function Specs({ chainSpecs, color }: SpecsProps) {
   );
 }
 
-function row(
+export function row(
   title: string,
   content: ReactElement | string,
   fullWidth = false,
